refactor notify
